+++ noodles-util/src/alignment/io/reader/inner.rs
@@ -5,6 +5,8 @@ use noodles_bgzf as bgzf;
 use noodles_cram as cram;
 use noodles_sam as sam;
 
+use crate::alignment::Record;
+
 pub(super) enum Inner<R> {
     Sam(sam::io::Reader<BufReader<R>>),
     SamGz(sam::io::Reader<bgzf::io::Reader<BufReader<R>>>),
@@ -24,6 +26,70 @@ where
             Inner::Bam(reader) => reader.read_header(),
             Inner::BamRaw(reader) => reader.read_header(),
             Inner::Cram(reader) => reader.get_mut().read_header(),
+        }
+    }
+
+    pub(super) fn read_record(
+        &mut self,
+        header: &sam::Header,
+        record: &mut Record,
+    ) -> io::Result<usize> {
+        match self {
+            Inner::Sam(reader) => {
+                if !matches!(record, Record::Sam(_)) {
+                    *record = Record::Sam(sam::Record::default());
+                }
+
+                if let Record::Sam(r) = record {
+                    reader.read_record(r)
+                } else {
+                    unreachable!();
+                }
+            }
+            Inner::SamGz(reader) => {
+                if !matches!(record, Record::Sam(_)) {
+                    *record = Record::Sam(sam::Record::default());
+                }
+
+                if let Record::Sam(r) = record {
+                    reader.read_record(r)
+                } else {
+                    unreachable!();
+                }
+            }
+            Inner::Bam(reader) => {
+                if !matches!(record, Record::Bam(_)) {
+                    *record = Record::Bam(bam::Record::default());
+                }
+
+                if let Record::Bam(r) = record {
+                    reader.read_record(r)
+                } else {
+                    unreachable!();
+                }
+            }
+            Inner::BamRaw(reader) => {
+                if !matches!(record, Record::Bam(_)) {
+                    *record = Record::Bam(bam::Record::default());
+                }
+
+                if let Record::Bam(r) = record {
+                    reader.read_record(r)
+                } else {
+                    unreachable!();
+                }
+            }
+            Inner::Cram(reader) => {
+                if !matches!(record, Record::Cram(_)) {
+                    *record = Record::Cram(sam::alignment::RecordBuf::default());
+                }
+
+                if let Record::Cram(r) = record {
+                    reader.read_record_buf(header, r)
+                } else {
+                    unreachable!();
+                }
+            }
         }
     }